# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Griffin webserver that serves, employee .",
        "title": "Griffin Web Server API Documentation",
        "contact": {
            "name": "Sang Il Bae",
            "email": "baesangil0906@gmail.com"
        },
        "version": "Document 1.0"
    },
    "host": "localhost:10433",
    "basePath": "/api/v0",
    "paths": {
        "/employType": {
            "get": {
                "description": "Whether the employer is full-time worker(permanent) or contract worker(freelance)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Employ type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enumerator (permanent or freelance)",
                        "name": "is_perma",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Single formed frequency, such as D(Day), W(Week)",
                        "name": "pay_freq",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.EMPLOY_TYPE"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Whether the employer is full-time worker(permanent) or contract worker(freelance)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add Employ type.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enumerator (permanent, freelance)",
                        "name": "is_perma",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Single formed frequency, such as D(Day), W(Week)",
                        "name": "pay_freq",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Not yet implemented",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Employ type.",
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_base.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_base.CommonResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "database (create / delete) (successful / failed)"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "ent.CRYPTO_CURRENCY": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CRYPTO_CURRENCYQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.CRYPTO_CURRENCYEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "source_id": {
                    "description": "SourceID holds the value of the \"source_id\" field.",
                    "type": "integer"
                },
                "ticker": {
                    "description": "Ticker holds the value of the \"ticker\" field.",
                    "type": "string"
                }
            }
        },
        "ent.CRYPTO_CURRENCYEdges": {
            "type": "object",
            "properties": {
                "currency_from_source": {
                    "description": "CurrencyFromSource holds the value of the currency_from_source edge.",
                    "$ref": "#/definitions/ent.CRYPTO_PRC_SOURCE"
                },
                "currency_of_employee": {
                    "description": "CurrencyOfEmployee holds the value of the currency_of_employee edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EMPLOYEE"
                    }
                },
                "currency_of_payment": {
                    "description": "CurrencyOfPayment holds the value of the currency_of_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT"
                    }
                },
                "currency_of_payment_history": {
                    "description": "CurrencyOfPaymentHistory holds the value of the currency_of_payment_history edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT_HISTORY"
                    }
                }
            }
        },
        "ent.CRYPTO_PRC_SOURCE": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CRYPTO_PRC_SOURCEQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.CRYPTO_PRC_SOURCEEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.CRYPTO_PRC_SOURCEEdges": {
            "type": "object",
            "properties": {
                "source_of_currency": {
                    "description": "SourceOfCurrency holds the value of the source_of_currency edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CRYPTO_CURRENCY"
                    }
                }
            }
        },
        "ent.EMPLOYEE": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "created_by": {
                    "description": "CreatedBy holds the value of the \"created_by\" field.",
                    "type": "string"
                },
                "crypto_currency_id": {
                    "description": "CryptoCurrencyID holds the value of the \"crypto_currency_id\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EMPLOYEEQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.EMPLOYEEEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "employ_type_id": {
                    "description": "EmployTypeID holds the value of the \"employ_type_id\" field.",
                    "type": "integer"
                },
                "employer_id": {
                    "description": "EmployerID holds the value of the \"employer_id\" field.",
                    "type": "integer"
                },
                "gid": {
                    "description": "Gid holds the value of the \"gid\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "payday": {
                    "description": "Payday holds the value of the \"payday\" field.",
                    "type": "string"
                },
                "payroll": {
                    "description": "Payroll holds the value of the \"payroll\" field.",
                    "type": "number"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "updated_by": {
                    "description": "UpdatedBy holds the value of the \"updated_by\" field.",
                    "type": "string"
                },
                "wallet": {
                    "description": "Wallet holds the value of the \"wallet\" field.",
                    "type": "string"
                },
                "work_ends": {
                    "description": "WorkEnds holds the value of the \"work_ends\" field.",
                    "type": "string"
                },
                "work_start": {
                    "description": "WorkStart holds the value of the \"work_start\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EMPLOYEEEdges": {
            "type": "object",
            "properties": {
                "employee_from_currency": {
                    "description": "EmployeeFromCurrency holds the value of the employee_from_currency edge.",
                    "$ref": "#/definitions/ent.CRYPTO_CURRENCY"
                },
                "employee_from_employ_type": {
                    "description": "EmployeeFromEmployType holds the value of the employee_from_employ_type edge.",
                    "$ref": "#/definitions/ent.EMPLOY_TYPE"
                },
                "employee_from_employer": {
                    "description": "EmployeeFromEmployer holds the value of the employee_from_employer edge.",
                    "$ref": "#/definitions/ent.EMPLOYER"
                },
                "employee_of_payment": {
                    "description": "EmployeeOfPayment holds the value of the employee_of_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT"
                    }
                },
                "employee_of_payment_history": {
                    "description": "EmployeeOfPaymentHistory holds the value of the employee_of_payment_history edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT_HISTORY"
                    }
                }
            }
        },
        "ent.EMPLOYER": {
            "type": "object",
            "properties": {
                "corp_email": {
                    "description": "CorpEmail holds the value of the \"corp_email\" field.",
                    "type": "string"
                },
                "corp_name": {
                    "description": "CorpName holds the value of the \"corp_name\" field.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "created_by": {
                    "description": "CreatedBy holds the value of the \"created_by\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EMPLOYERQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.EMPLOYEREdges"
                },
                "gid": {
                    "description": "Gid holds the value of the \"gid\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "updated_by": {
                    "description": "UpdatedBy holds the value of the \"updated_by\" field.",
                    "type": "string"
                },
                "username": {
                    "description": "Username holds the value of the \"username\" field.",
                    "type": "string"
                },
                "wallet": {
                    "description": "Wallet holds the value of the \"wallet\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EMPLOYEREdges": {
            "type": "object",
            "properties": {
                "employer_of_employee": {
                    "description": "EmployerOfEmployee holds the value of the employer_of_employee edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EMPLOYEE"
                    }
                },
                "employer_of_payment": {
                    "description": "EmployerOfPayment holds the value of the employer_of_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT"
                    }
                },
                "employer_of_payment_history": {
                    "description": "EmployerOfPaymentHistory holds the value of the employer_of_payment_history edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PAYMENT_HISTORY"
                    }
                }
            }
        },
        "ent.EMPLOY_TYPE": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EMPLOY_TYPEQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.EMPLOY_TYPEEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "is_permanent": {
                    "description": "IsPermanent holds the value of the \"is_permanent\" field.",
                    "type": "string"
                },
                "pay_freq": {
                    "description": "PayFreq holds the value of the \"pay_freq\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EMPLOY_TYPEEdges": {
            "type": "object",
            "properties": {
                "employ_type_of_employee": {
                    "description": "EmployTypeOfEmployee holds the value of the employ_type_of_employee edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EMPLOYEE"
                    }
                }
            }
        },
        "ent.PAYMENT": {
            "type": "object",
            "properties": {
                "crypto_currency_id": {
                    "description": "CryptoCurrencyID holds the value of the \"crypto_currency_id\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PAYMENTQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.PAYMENTEdges"
                },
                "employee_id": {
                    "description": "EmployeeID holds the value of the \"employee_id\" field.",
                    "type": "integer"
                },
                "employer_id": {
                    "description": "EmployerID holds the value of the \"employer_id\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "payment_amount": {
                    "description": "PaymentAmount holds the value of the \"payment_amount\" field.",
                    "type": "number"
                },
                "payment_executed": {
                    "description": "PaymentExecuted holds the value of the \"payment_executed\" field.",
                    "type": "string"
                },
                "payment_scheduled": {
                    "description": "PaymentScheduled holds the value of the \"payment_scheduled\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PAYMENTEdges": {
            "type": "object",
            "properties": {
                "payment_from_currency": {
                    "description": "PaymentFromCurrency holds the value of the payment_from_currency edge.",
                    "$ref": "#/definitions/ent.CRYPTO_CURRENCY"
                },
                "payment_from_employee": {
                    "description": "PaymentFromEmployee holds the value of the payment_from_employee edge.",
                    "$ref": "#/definitions/ent.EMPLOYEE"
                },
                "payment_from_employer": {
                    "description": "PaymentFromEmployer holds the value of the payment_from_employer edge.",
                    "$ref": "#/definitions/ent.EMPLOYER"
                }
            }
        },
        "ent.PAYMENT_HISTORY": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "number"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "created_by": {
                    "description": "CreatedBy holds the value of the \"created_by\" field.",
                    "type": "string"
                },
                "currency_id": {
                    "description": "CurrencyID holds the value of the \"currency_id\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PAYMENT_HISTORYQuery when eager-loading is set.",
                    "$ref": "#/definitions/ent.PAYMENT_HISTORYEdges"
                },
                "employee_id": {
                    "description": "EmployeeID holds the value of the \"employee_id\" field.",
                    "type": "integer"
                },
                "employer_id": {
                    "description": "EmployerID holds the value of the \"employer_id\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PAYMENT_HISTORYEdges": {
            "type": "object",
            "properties": {
                "payment_history_from_currency_id": {
                    "description": "PaymentHistoryFromCurrencyID holds the value of the payment_history_from_currency_id edge.",
                    "$ref": "#/definitions/ent.CRYPTO_CURRENCY"
                },
                "payment_history_from_employee": {
                    "description": "PaymentHistoryFromEmployee holds the value of the payment_history_from_employee edge.",
                    "$ref": "#/definitions/ent.EMPLOYEE"
                },
                "payment_history_from_employer": {
                    "description": "PaymentHistoryFromEmployer holds the value of the payment_history_from_employer edge.",
                    "$ref": "#/definitions/ent.EMPLOYER"
                }
            }
        },
...
    }
}